Make the CSV file work again with all datasets.
