+++ src/omnipose.py
@@ -39,3 +39,4 @@ if __name__ == '__main__':
     root = "./data/raw"
     dataset = OmniPoseSet()
     dataset.unpack(root)
+    dataset.crosscheck(root)
